% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast-R0-full.R
\name{BEAST_R0_Full}
\alias{BEAST_R0_Full}
\title{BEAST_R0_Full}
\usage{
BEAST_R0_Full(N = 1e+05, R0 = 2, Tg = 2, t.cont = TRUE,
  real.time = FALSE, scale = 1, size.reps = c(0, 10, 1, 0.1),
  mu = 1.57e-05, root.path = NULL, seq.length = 1600, model = "HKY85",
  evolution.model = "within.host", kappa = 2, outgroup = FALSE,
  missed.generations = 0, n.seq.samples = 50, sample.reps = 10,
  tree.reps = 20, sigma.reps = c(1, 5, 9, 13, 18), oldestboolean = FALSE,
  template.xml = "C:/Users/Oliver/GoogleDrive/AcademicWork/Imperial/git/evoflution/inst/extdata/template.xml",
  dir = "C:/Users/Oliver/GoogleDrive/AcademicWork/Imperial/git/evoflution/Results/Within_Host_No_Outgroup",
  network.path = "\\\\\\\\fi--san02.dide.ic.ac.uk\\\\homes\\\\olw13\\\\Data\\\\")
}
\arguments{
\item{N}{Target infected population size.  Default = 1e5}

\item{R0}{basic reproduction number. Default = 2}

\item{Tg}{mean generation time. Default = 2.6 days}

\item{t.cont}{Boolean dictating whether gamma process is used for generation time distribution. Default = TRUE}

\item{real.time}{Boolean dictating. whether an extra generation is simulated producing extra for real.time consideration.Default = FALSE}

\item{scale}{rgamma scale parameter. Default = 1}

\item{size.reps}{Vector of different size parameters in negative binomial offspring distribution. Default=c(0,10,1,0.1)}

\item{mu}{Clock Rate. Default=1e-4}

\item{root.path}{File path of saveRDS(root). Default = NULL}

\item{seq.length}{Sequence length. Will be overwritten accordingly if root.path given. Default = 1701}

\item{model}{Nucleic subsitution model. Can be one of "JC69", "K80", "HKY85" (Default)}

\item{evolution.model}{String variable detailing evolutionary model. Can be any of "generational", "within.host" and "parallel"}

\item{kappa}{Tv/Ts ration. Default = 2}

\item{missed.generations}{Number of intiial generations not considered for sampling. Default = 0}

\item{n.seq.samples}{number of sequences to be sampled. Default = 50}

\item{sample.reps}{number of tree subsamples under each scheme. Default = 10}

\item{tree.reps}{number of unique tree simulations conducted under each scheme. Default = 20}

\item{sigma.reps}{variables total. Default = c(1,5,9,13,18)}

\item{oldestboolean}{boolean if oldest sequence not root is included. Default=TRUE}

\item{template.xml}{path for template xml. Default is a strict molecular clock, exponential coalescent, HKY85+G4 xml
found within the data folder called test.xml}

\item{dir}{Local Directory where simulations tree and nexus files are stored. Defaults specific to author.}

\item{network.path}{Network Directory which a cluster can see to access the BEAST xml files. Defaults specific to author.}

\item{size.reps}{- Vector of offspring distribution parameters. Default=c(0,10,1,0.1)}

\item{outgoup}{Boolean concerning whether outgroup is simulated. Default = FALSE}
}
\value{
Returns a data frame of collated statstics
}
\description{
Main simulation function to investigate the effects of sampling strategy against TMRCA/Clock Rate Bias. As opposed
to \code{BEAST_R0_Check}, this function considers all sampling schemes.
}
\details{
\code{BEAST_R0_Full} proceeds by simulating a tree in a local directory, sampling from it and placing the nexus samples
within the local directory, before then creating the BEAST2.3xml file in a parallel network directory, and
then running the BEAST analysis using the batch collation file.

Highly prescriptive function declaration, which considers all the simulation parameters needed, the range of
sigma parameters, the BEAST template required, directories etc.
}

