% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast-R0-network-tracer-accumlation.R
\name{BEAST_R0_Network_Pull_Tracer}
\alias{BEAST_R0_Network_Pull_Tracer}
\title{BEAST_R0_Network_Pull_Tracer}
\usage{
BEAST_R0_Network_Pull_Tracer(sampling.reps = 20, rep.name = "bs",
  outgroup = FALSE, tree.reps = 10, vars = 5, var.R0 = 0, sigma = 1,
  N.v = c(rep(1e+05, tree.reps)), mu.v = c(rep(1.57e-05, tree.reps)),
  Tg.v = c(rep(2.6, tree.reps)), R0.v = c(rep(2, tree.reps)),
  network.root.dir, local.root.dir)
}
\arguments{
\item{sampling.reps}{Sampling replicates total. Default = 20}

\item{rep.name}{The character name used for each sampling rep, i.e. Unifrom, Random, Oldest_Bar_One}

\item{outgroup}{Boolean describing if an outgroup was present}

\item{tree.reps}{number of subdirectories to be scanned}

\item{vars}{Tracer variables t collect in total. Default = 5}

\item{var.R0}{Size parameter for offspring distribtuion. Default = 0}

\item{sigma}{value Sigma value used}

\item{N.v}{column vector of N. Default: N.v=c(rep(100000,tree.reps)),}

\item{mu.v}{column vector of mutation rates in directory. Default: mu.v=c(rep(1.57e-05,tree.reps)),}

\item{Tg.v}{column vector of generation times in directory. Default: Tg.v=c(rep(2.6,tree.reps)),}

\item{R0.v}{column vector of R0 in directory. Default: R0.v=c(rep(2,tree.reps)),}

\item{network.root.dir}{network root directory}

\item{local.root.dir}{local root directory}
}
\value{
Returns a data frame of collated statstics
}
\description{
Summary Statistics are clock rate, growthRate, ePop and TMRCA.
}

