% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias-plot.R
\name{Bias_Plot}
\alias{Bias_Plot}
\title{Bias plotting}
\usage{
Bias_Plot(dataset)
}
\arguments{
\item{dataset}{Output of pipeline}
}
\value{
plots 4 x 1 boxplots and returns the ggplots
}
\description{
This function takes the output of a pipeline and plots posterior parameter estimate biases
}

