% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKY85.R
\name{HKY85}
\alias{HKY85}
\title{HKY85 Substiution Model}
\usage{
HKY85(t, mu, kappa, pi, length = NULL)
}
\arguments{
\item{t}{Branch length, i.e. specific generation time}

\item{mu}{Mutation rate}

\item{kappa}{ransition/transversion rate ratio}

\item{pi}{Base Frequencies}

\item{length}{Length of sequence that was used in creating the base frequencies.
If provided the output will instead be the probability of mutation occuring}
}
\value{
This function returns a \code{dgeMatrix} containing the transition matrix P, except if length is provided
}
\description{
\code{HKY85} creates the rate matrix of the HKY85 model before calculating the
related transition matrix, P, using the mutation (mu), branch length/time (t) and transition
transversion ratio (kappah)
}

