% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JC69.R
\name{JC69}
\alias{JC69}
\title{JC69 Substiution Model}
\usage{
JC69(t, mu, length = NULL)
}
\arguments{
\item{t}{- branch length, i.e. specific generation time}

\item{mu}{- mutation rate}

\item{length}{Length of sequence that was used in creating the base frequencies.
If provided the output will instead be the probability of mutation occuring}
}
\value{
This function returns a \code{dgeMatrix} containing the transition matrix P, except if length is provided
}
\description{
\code{JC69} creates the rate matrix of the JC69 model before calculating the
related transition matrix, P, using the mutation and branch length/time from parent
variables.
}

