% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K80.R
\name{K80}
\alias{K80}
\title{K80 Substiution Model}
\usage{
K80(t, mu, kappa, pi = c(0.25, 0.25, 0.25, 0.25), length = NULL)
}
\arguments{
\item{t}{Branch length, i.e. specific generation time}

\item{mu}{Mutation rate}

\item{kappa}{ransition/transversion rate ratio}

\item{pi}{Base Frequencies}

\item{length}{Length of sequence that was used in creating the base frequencies.
If provided the output will instead be the probability of mutation occuring}
}
\value{
This function returns a \code{dgeMatrix} containing the transition matrix P, except if length is provided
}
\description{
\code{K80} This function creates the rate matrix of the K80 model before calculating the
related transition matrix, P, using the mutation (mu), branch length/time (t) and transition
transversion ratio (kappah)
}

