% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_xml.R
\name{List_To_Xml}
\alias{List_To_Xml}
\title{List to XML}
\usage{
List_To_Xml(item, tag)
}
\arguments{
\item{tag}{- the opening tag required for the xml creation, e.g. "beast"}

\item{list}{- list or list of lists pertaining to the xml}
}
\value{
xmlNode
}
\description{
Convert List to XML
}
\details{
Can convert list or other object to an xml object using xmlNode
}

