% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{New_sampling}
\alias{New_sampling}
\title{Forwards genetic sequence simulator}
\usage{
New_sampling(sim.out, seq.out, file, n.seq.samples = 50, bs.label = "rep",
  sigma = 1, n.bs = 50, random = TRUE, oldestboolean = TRUE,
  outgroup = TRUE, discretised = FALSE, proportional = FALSE,
  t.cont = TRUE, missed.generations = 0)
}
\arguments{
\item{sim.out}{Output of \code{Infection_Sim}}

\item{seq.out}{Output of \code{Sequencing}}

\item{file}{Directory path where the nexus file of samples are to be saved to.}

\item{n.seq.samples}{Number of sequences to be sampled in total}

\item{bs.label}{String used to be added on to the created nex files. Default = "bs"}

\item{sigma}{The interval parameter, where the interval of sequences is equal to 1/sigma. Default = 1, i.e. all sequences}

\item{n.bs}{Number of bootstrap samples to create. Default = 50}

\item{oldestboolean}{Boolean if the oldest sample is included (TRUE). Default=TRUE}

\item{outgroup}{Boolean if the outgroup  is included (TRUE). Default=FALSE}

\item{discretised}{Boolean if discretised sampling occurs. Default = FALSE}

\item{proportional}{Boolean if proportional sampling occurs. Default = FALSE}

\item{t.cont}{Boolean if the simulation occurs in discrete (FALSE) or continuous time (TRUE), in turn directing
whether decimilisation is needed in uniform sampling strategies.}

\item{missed.generations}{Number of intiial generations not considered for sampling. Default = 0}
}
\value{
save a nexus file of dated sequences
}
\description{
\code{New_sampling} uses the output of \code{Infection_Sim} and \code{Sequencing}
to sample a selection of sequences and write the annotated output top file as a fasta
}
\note{
If the combination of prop.times and extant==FALSE creates a time window that does not include all clades, and all.clades
== TRUE, only all the possible clades within the time window will be sampled.
}

