% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nex-to-beast-xml.R
\name{Nex_2_BEASTxml}
\alias{Nex_2_BEASTxml}
\title{Take nex and turn into BEAST 2.3 xml file}
\usage{
Nex_2_BEASTxml(template.xml, nex, file, rep.str, cont = FALSE,
  relaxed = FALSE)
}
\arguments{
\item{template.xml}{- the template xml already parsed as an xml object via xmlParse > xmlToList}

\item{nex}{- the nex file path to be used to create the new beast.xml}

\item{file}{- the full file path where the created beast xml is to be written to}

\item{rep.str}{- the id string that is to replace the id string from the template xml, usually replicate#}

\item{cont}{- boolean concerning whether the simulation is time explicit or not. Default = FALSE}
}
\value{
written BEAST.xml
}
\description{
This function takes a generated nex file, and creates a related BEAST 2.3 xml
that uses another BEAST xml file as a template.
}
\note{
Function has been written to be used ideally within a loop of using sample_build directly afterwards,
and in this instance it makes more sense to predetermine the template.xml and template.txt. Hence the
template.path is default null, and function written intending user to allocated the xml and txt templates.
}

