% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{Sequencing}
\alias{Sequencing}
\title{Forwards genetic sequence simulator}
\usage{
Sequencing(sim.out, seq.length = 1701, mu = 1.57e-05, Tg = 2.6,
  model = "HKY85", kappa = 2, root.path = NULL,
  evolution.model = "generational", outgroup = TRUE)
}
\arguments{
\item{sim.out}{Output of \code{Infection_Sim}}

\item{seq.length}{Sequence length. Will be overwritten accordingly if root.path given. Default = 1701}

\item{mu}{Daily mutation rate. Default = 1.57e-05 sub / site / day}

\item{Tg}{mean generation time. Default = 2.6 days}

\item{model}{Nucleic subsitution model. Can be one of "JC69", "K80", "HKY85" (Default)}

\item{kappa}{Tv/Ts ration. Default = 2}

\item{root.path}{File path of saveRDS(root). Default = NULL}

\item{evolution.model}{String variable detailing evolutionary model. Can be any of "generational", "within.host" and "parallel"}
}
\value{
Returns a list of 4 lists
}
\description{
\code{Sequencing} uses the output of \code{Infection_Sim} to simualate the evolutionary
process of the seqeunce from the root forwards
}

