% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-addon.R
\name{Tracer_Addon}
\alias{Tracer_Addon}
\title{Tracer Addon Function}
\usage{

  Tracer_Addon(cmd = paste("\\"C:\\\\Users\\\\Oliver\\\\GoogleDrive\\\\AcademicWork\\\\Imperial\\\\O15-12\\\\BEAST\\\\BEAST v1.8.2\\\\bin\\\\loganalyser.cmd\\""),
  file.dir, log.dir, reps, replicate.name = "replicate")
}
\arguments{
\item{cmd}{shellscript path name for loganalyser.cmd. Tracer Addon uses the internal
function system(), so path must be written accordingly.}

\item{file.dir}{directory where the sumamry.txt file will be written to}

\item{log.dir}{directory from which the log files are to be read from}

\item{reps}{number of reps within the log.dir}
}
\value{
writes summary.txt file
}
\description{
This function links with Tracer_Accumulation by creating the summary.txt file
needed, by loading the log files as needed and first writing it.
}
\details{
Tree Statistics pulled are internally coded to be the posterior, clock rate,
growthRate, ePop and TMRCA.

The summary statistics pulled are internally coded to be the mean, standard error,
median, HPD 95% upper and lower estimate and effective sample size.

The output is written to file for further use
}

